New implementation with typer
